move UpdateLayoutReferences into a helper
