+++ src/Designer/backend/src/Designer/EventHandlers/LayoutSetDeleted/LayoutSetDeletedLayoutsHandler.cs
@@ -13,7 +13,7 @@ namespace Altinn.Studio.Designer.EventHandlers.LayoutSetDeleted;
 
 public class LayoutSetDeletedLayoutsHandler(
     IFileSyncHandlerExecutor fileSyncHandlerExecutor,
-    IUiFoldersService uiFoldersService
+    ILayoutReferenceUpdater layoutReferenceUpdater
 ) : INotificationHandler<LayoutSetDeletedEvent>
 {
     public async Task Handle(LayoutSetDeletedEvent notification, CancellationToken cancellationToken)
@@ -28,7 +28,7 @@ public class LayoutSetDeletedLayoutsHandler(
                 [
                     new Reference(ReferenceType.LayoutSet, notification.LayoutSetName, notification.LayoutSetName),
                 ];
-                return await uiFoldersService.UpdateLayoutReferences(
+                return await layoutReferenceUpdater.UpdateLayoutReferences(
                     notification.EditingContext,
                     referencesToDelete,
                     cancellationToken
