+++ src/Designer/backend/src/Designer/Services/Interfaces/IAppDevelopmentService.cs
@@ -304,4 +304,16 @@ public interface IAppDevelopmentService
         ValidationOnNavigation validationOnNavigation,
         CancellationToken cancellationToken
     );
+
+    /// <summary>
+    /// Update layout references
+    /// </summary>
+    /// <param name="altinnRepoEditingContext">An <see cref="AltinnRepoEditingContext"/>.</param>
+    /// <param name="referencesToUpdate">The references to update.</param>
+    /// <param name="cancellationToken">An <see cref="CancellationToken"/> that observes if operation is cancelled.</param>
+    public Task<bool> UpdateLayoutReferences(
+        AltinnRepoEditingContext altinnRepoEditingContext,
+        List<Reference> referencesToUpdate,
+        CancellationToken cancellationToken
+    );
 }
